adding product repository and service,for registering and searching products

+++ app/src/main/java/com/nozama/app/controller/ProductController.java
@@ -5,11 +5,11 @@ import com.nozama.app.dto.ProductResponse;
 import com.nozama.app.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -27,8 +27,8 @@ public class ProductController {
   public ResponseEntity<List<ProductResponse>> searchProducts(
           @RequestParam(required = false) String search,
           @RequestParam(required = false) String category,
-          @RequestParam(required = false) double minPrice,
-          @RequestParam(required = false) double maxPrice
+          @RequestParam(required = false) BigDecimal minPrice,
+          @RequestParam(required = false) BigDecimal maxPrice
   ) {
     return ResponseEntity.ok(productService.searchProducts(search, category, minPrice, maxPrice));
   }
